requirements.txt
fastapi==0.135.1
langchain==1.2.12
langchain-community==0.4.1
langchain-core==1.2.19
langchain-google-genai==4.2.1
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.0
pinecone==7.3.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pypdf==6.10.2
python-dotenv==1.2.2
SQLAlchemy==2.0.48
uvicorn==0.42.0
